Unit tests passing before adding more complex ones
