--- runtime/test_artifacts/genesis.go
@@ -1,11 +1,0 @@
-package test_artifacts
-
-import "fmt"
-
-const (
-	genesisStatePostfix = "_genesis_state"
-)
-
-func GetGenesisFileName(moduleName string) string {
-	return fmt.Sprintf("%s%s", moduleName, genesisStatePostfix)
-}
