+++ shared/core/types/proto/actor.proto
@@ -20,16 +20,17 @@ enum StakeStatus {
 }
 
 // TODO(#555): Investigate ways of having actor specific params that are not part of the shared struct.
+//             Potentially having a separate struct for each actor type and a shared base.
 message Actor {
   ActorType actor_type = 1;
   string address = 2;
   string public_key = 3;
-  repeated string chains = 4; // Not applicable to `Validator` actors
+  repeated string chains = 4; // NB: Not applicable `Validator` actors
   // proto-gen-c does not support `go_name` at the time of writing resulting
   // in the output go field being snakeCase: ServiceUrl (https://github.com/golang/protobuf/issues/555)
   string service_url = 5; // Not applicable to `Application` actors
   string staked_amount = 6;
-  int64 paused_height = 7;
-  int64 unstaking_height = 8;
-  string output = 9;
+  int64 paused_height = 7; // TECHDEBT: Revisit this parameter and see if it can be removed for simplification purposes.
+  int64 unstaking_height = 8; // TECHDEBT: Revisit this parameter and see if it can be removed for simplification purposes.
+  string output = 9; // TECHDEBT: Revisit custodial / non-custodial flows (e.g. what if we want multiple outsputs for business purposes)
 }
