+++ utility/session_test.go
@@ -3,30 +3,205 @@ package utility
 import (
 	"testing"
 
+	"github.com/pokt-network/pocket/runtime/test_artifacts"
 	"github.com/pokt-network/pocket/shared/messaging"
 	"github.com/pokt-network/pocket/utility/types"
 	"github.com/stretchr/testify/require"
 )
 
-func TestSession_NewSession_SimpleCase(t *testing.T) {
-	height := int64(1)
-	relayChain := "0001"
-	geoZone := "unused_geo"
+// TECHDEBT: Geozones are not current implemented, used or tested
 
-	runtimeCfg, utilityMod, _ := prepareEnvironment(t, 5, 1, 1, 1)
+// One fishermen and one servicer happy base case
+func TestSession_NewSession_BaseCase(t *testing.T) {
+	// Test parameters
+	height := int64(1)
+	relayChain := test_artifacts.DefaultChains[0]
+	geoZone := "unused_geo"
+	numFishermen := 1
+	numServicers := 1
+	expectedSessionId := "3545185ff1519bf7706ec8f828d16525830d3c0dcc2425c40db597ee6b67b8bc" // needs to be manually updated if business logic changes
+
+	runtimeCfg, utilityMod, _ := prepareEnvironment(t, 5, numServicers, 1, numFishermen)
+
+	// Sanity check genesis
 	require.Len(t, runtimeCfg.GetGenesis().Applications, 1)
 	app := runtimeCfg.GetGenesis().Applications[0]
+	require.Len(t, runtimeCfg.GetGenesis().Fishermen, 1)
+	fish := runtimeCfg.GetGenesis().Fishermen[0]
+	require.Len(t, runtimeCfg.GetGenesis().Servicers, 1)
+	servicer := runtimeCfg.GetGenesis().Servicers[0]
 
 	session, err := utilityMod.GetSession(app.Address, height, relayChain, geoZone)
 	require.NoError(t, err)
-	require.Equal(t, "3545185ff1519bf7706ec8f828d16525830d3c0dcc2425c40db597ee6b67b8bc", session.Id)
+	require.Equal(t, expectedSessionId, session.Id)
 	require.Equal(t, height, session.Height)
 	require.Equal(t, relayChain, session.RelayChain)
 	require.Equal(t, geoZone, session.GeoZone)
-	require.Equal(t, session.Application.Address, app.Address)
-	require.Equal(t, "c7832263600476fd6ff4c5cb0a86080d0e5f48b2", session.Servicers[0].Address)
-	require.Equal(t, "a6e7b6810df8120580f2a81710e228f454f99c97", session.Fishermen[0].Address)
+	require.Equal(t, app.Address, session.Application.Address)
+	require.Len(t, session.Servicers, numServicers)
+	require.Equal(t, servicer.Address, session.Servicers[0].Address)
+	require.Len(t, session.Fishermen, numFishermen)
+	require.Equal(t, fish.Address, session.Fishermen[0].Address)
 }
+
+func TestSession_ServicersAndFishermanCounts_TotalAvailability(t *testing.T) {
+	numServicers := 100
+	numFishermen := 100
+	runtimeCfg, utilityMod, persistenceMod := prepareEnvironment(t, 5, numServicers, 1, numFishermen)
+
+	app := runtimeCfg.GetGenesis().Applications[0]
+	relayChain := test_artifacts.DefaultChains[0]
+	geoZone := "unused_geo"
+
+	tests := []struct {
+		name                   string
+		numServicersPerSession int64
+		numFishermanPerSession int64
+		wantServicerCount      int
+		wantFishermanCount     int
+	}{
+		{
+			name:                   "more actors per session than available in network",
+			numServicersPerSession: int64(numServicers) * 10,
+			numFishermanPerSession: int64(numFishermen) * 10,
+			wantServicerCount:      numServicers,
+			wantFishermanCount:     numFishermen,
+		},
+		{
+			name:                   "less actors per session than available in network",
+			numServicersPerSession: int64(numServicers) / 2,
+			numFishermanPerSession: int64(numFishermen) / 2,
+			wantServicerCount:      numServicers / 2,
+			wantFishermanCount:     numFishermen / 2,
+		},
+		{
+			name:                   "same number of actors per session as available in network",
+			numServicersPerSession: int64(numServicers),
+			numFishermanPerSession: int64(numFishermen),
+			wantServicerCount:      numServicers,
+			wantFishermanCount:     numFishermen,
+		},
+		{
+			name:                   "more than enough servicers but not enough fishermen",
+			numServicersPerSession: int64(numServicers) / 2,
+			numFishermanPerSession: int64(numFishermen) * 10,
+			wantServicerCount:      numServicers / 2,
+			wantFishermanCount:     numFishermen,
+		},
+		{
+			name:                   "more than enough fishermen but not enough servicers",
+			numServicersPerSession: int64(numServicers) * 10,
+			numFishermanPerSession: int64(numFishermen) / 2,
+			wantServicerCount:      numServicers,
+			wantFishermanCount:     numFishermen / 2,
+		},
+	}
+
+	updateParamsHeight := int64(1)
+	querySessionHeight := int64(2)
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := persistenceMod.HandleDebugMessage(&messaging.DebugMessage{
+				Action:  messaging.DebugMessageAction_DEBUG_PERSISTENCE_RESET_TO_GENESIS,
+				Message: nil,
+			})
+			require.NoError(t, err)
+
+			writeCtx, err := persistenceMod.NewRWContext(updateParamsHeight)
+			require.NoError(t, err)
+			defer writeCtx.Release()
+
+			err = writeCtx.SetParam(types.ServicersPerSessionParamName, tt.numServicersPerSession)
+			require.NoError(t, err)
+			err = writeCtx.SetParam(types.FishermanPerSessionParamName, tt.numFishermanPerSession)
+			require.NoError(t, err)
+			err = writeCtx.Commit([]byte("empty_proposed_addr"), []byte("empty_quorum_cert"))
+			require.NoError(t, err)
+
+			session, err := utilityMod.GetSession(app.Address, querySessionHeight, relayChain, geoZone)
+			require.NoError(t, err)
+			require.Equal(t, tt.wantServicerCount, len(session.Servicers))
+			require.Equal(t, tt.wantFishermanCount, len(session.Fishermen))
+		})
+	}
+}
+
+func TestSession_ServicersAndFishermanCounts_ChainAvailability(t *testing.T) {
+	// numServicers := 100
+	// numFishermen := 100
+	// runtimeCfg, utilityMod, persistenceMod := prepareEnvironment(t, 5, numServicers, 1, numFishermen)
+
+	// app := runtimeCfg.GetGenesis().Applications[0]
+	// relayChain := test_artifacts.DefaultChains[0]
+	// geoZone := "unused_geo"
+
+	// tests := []struct {
+	// 	name                   string
+	// 	numServicersPerSession int64
+	// 	numFishermanPerSession int64
+	// 	wantServicerCount      int
+	// 	wantFishermanCount     int
+	// }{
+	// 	{
+	// 		name:                   "more actors per session than available in network",
+	// 		numServicersPerSession: int64(numServicers) * 10,
+	// 		numFishermanPerSession: int64(numFishermen) * 10,
+	// 		wantServicerCount:      numServicers,
+	// 		wantFishermanCount:     numFishermen,
+	// 	},
+	// 	{
+	// 		name:                   "less actors per session than available in network",
+	// 		numServicersPerSession: int64(numServicers) / 2,
+	// 		numFishermanPerSession: int64(numFishermen) / 2,
+	// 		wantServicerCount:      numServicers / 2,
+	// 		wantFishermanCount:     numFishermen / 2,
+	// 	},
+	// 	{
+	// 		name:                   "same number of actors per session as available in network",
+	// 		numServicersPerSession: int64(numServicers),
+	// 		numFishermanPerSession: int64(numFishermen),
+	// 		wantServicerCount:      numServicers,
+	// 		wantFishermanCount:     numFishermen,
+	// 	},
+	// }
+
+	// updateParamsHeight := int64(1)
+	// querySessionHeight := int64(2)
+
+	// for _, tt := range tests {
+	// 	t.Run(tt.name, func(t *testing.T) {
+	// 		err := persistenceMod.HandleDebugMessage(&messaging.DebugMessage{
+	// 			Action:  messaging.DebugMessageAction_DEBUG_PERSISTENCE_RESET_TO_GENESIS,
+	// 			Message: nil,
+	// 		})
+	// 		require.NoError(t, err)
+
+	// 		writeCtx, err := persistenceMod.NewRWContext(updateParamsHeight)
+	// 		require.NoError(t, err)
+	// 		defer writeCtx.Release()
+
+	// 		err = writeCtx.SetParam(types.ServicersPerSessionParamName, tt.numServicersPerSession)
+	// 		require.NoError(t, err)
+	// 		err = writeCtx.SetParam(types.FishermanPerSessionParamName, tt.numFishermanPerSession)
+	// 		require.NoError(t, err)
+	// 		err = writeCtx.Commit([]byte("empty_proposed_addr"), []byte("empty_quorum_cert"))
+	// 		require.NoError(t, err)
+
+	// 		session, err := utilityMod.GetSession(app.Address, querySessionHeight, relayChain, geoZone)
+	// 		require.NoError(t, err)
+	// 		require.Equal(t, tt.wantServicerCount, len(session.Servicers))
+	// 		require.Equal(t, tt.wantFishermanCount, len(session.Fishermen))
+	// 	})
+	// }
+}
+
+// Not enough servicers in region
+// Not enough fisherman in region
+// Not enough servicers per chain
+// Not enough fisherman per chain
+
+// func TestSession_NewSession_BaseCase(t *testing.T) {
 
 // dispatching session in the future
 // dispatching session in the past
@@ -46,83 +221,6 @@ func TestSession_NewSession_SimpleCase(t *testing.T) {
 // invalid geozone
 // unused geozone
 // non-existent geozone
-
-func TestSession_ServicersAndFishermanCount(t *testing.T) {
-	numServicers := 100
-	numFishermen := 100
-	// Prepare an environment with lots of servicers and fisherman
-	runtimeCfg, utilityMod, persistenceMod := prepareEnvironment(t, 5, numServicers, 1, numFishermen)
-
-	app := runtimeCfg.GetGenesis().Applications[0]
-
-	// height := int64(1)
-	relayChain := "0001"
-	geoZone := "unused_geo"
-
-	// defer writeCtx.Release()
-
-	tests := []struct {
-		name                   string
-		numServicersPerSession int64
-		numFishermanPerSession int64
-		wantServicerCount      int
-		wantFishermanCount     int
-	}{
-		{
-			name:                   "not enough actors to choose from (> 100)",
-			numServicersPerSession: int64(numServicers) * 10,
-			numFishermanPerSession: int64(numFishermen) * 10,
-			wantServicerCount:      numServicers,
-			wantFishermanCount:     numFishermen,
-		},
-		{
-			name:                   "too many actors to choose from (< 100)",
-			numServicersPerSession: int64(numServicers) / 2,
-			numFishermanPerSession: int64(numFishermen) / 2,
-			wantServicerCount:      numServicers / 2,
-			wantFishermanCount:     numFishermen / 2,
-		},
-		{
-			name:                   "same number of servicers and fisherman",
-			numServicersPerSession: int64(numServicers),
-			numFishermanPerSession: int64(numFishermen),
-			wantServicerCount:      numServicers,
-			wantFishermanCount:     numFishermen,
-		},
-		// Not enough servicers in region
-		// Not enough fisherman in region
-		// Not enough servicers per chain
-		// Not enough fisherman per chain
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := persistenceMod.HandleDebugMessage(&messaging.DebugMessage{
-				Action:  messaging.DebugMessageAction_DEBUG_PERSISTENCE_RESET_TO_GENESIS,
-				Message: nil,
-			})
-			require.NoError(t, err)
-			writeCtx, err := persistenceMod.NewRWContext(1)
-			require.NoError(t, err)
-			defer writeCtx.Release()
-
-			err = writeCtx.SetParam(types.ServicersPerSessionParamName, tt.numServicersPerSession)
-			require.NoError(t, err)
-			writeCtx.SetParam(types.FishermanPerSessionParamName, tt.numFishermanPerSession)
-			require.NoError(t, err)
-
-			err = writeCtx.Commit([]byte(""), []byte(""))
-			require.NoError(t, err)
-
-			session, err := utilityMod.GetSession(app.Address, 2, relayChain, geoZone)
-
-			// require.NoError(t, writeCtx.Commit([]byte(""), []byte("")))
-			require.NoError(t, err)
-			require.Equal(t, tt.wantServicerCount, len(session.Servicers))
-			require.Equal(t, tt.wantFishermanCount, len(session.Fishermen))
-		})
-	}
-}
 
 // generate session id
